Moved examples function to setup.py to avoid dependencies install
