+++ holoviews/util/command.py
@@ -24,7 +24,7 @@ except:
     from holoviews.ipython.preprocessors import OptsMagicProcessor, OutputMagicProcessor
     from holoviews.ipython.preprocessors import StripMagicsProcessor
 
-from . import examples
+from ..setup import examples
 
 
 def main():
